+++ Runtime/Utilities/Singleton/PersistentSingleton.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 using NekoLib.Extensions;
+using NekoLib.ColorPalette;
 
 namespace NekoLib.Singleton
 {
@@ -11,18 +12,18 @@ namespace NekoLib.Singleton
     public abstract class PersistentSingleton<T> : MonoBehaviour where T : MonoBehaviour
     {
         private static T s_instance;
-        private static readonly object _lock = new();
+        private static readonly object s_lock = new();
         private static bool s_applicationIsQuitting = false;
 
         public static T Instance
         {
             get
             {
-                lock (_lock)
+                lock (s_lock)
                 {
                     if (s_applicationIsQuitting)
                     {
-                        Debug.LogWarning($"Instance {typeof(T).Name.Colorize(Color.magenta)} already destroyed on application quit. Won't create again - returning null.");
+                        Debug.LogWarning($"Instance {typeof(T).Name.Colorize(Palette.GoldenAmber)} already destroyed on application quit. Won't create again - returning null.");
                         return null;
                     }
 
@@ -34,7 +35,7 @@ namespace NekoLib.Singleton
                         };
 
                         s_instance = obj.AddComponent<T>();
-                        Debug.Log($"Create a new singleton of type {typeof(T).Name.Colorize(Color.magenta)}.");
+                        Debug.Log($"Create a new singleton of type {typeof(T).Name.Colorize(Palette.Lavender)}.");
                     }
                 }
 
@@ -52,7 +53,7 @@ namespace NekoLib.Singleton
             {
                 if (s_instance != this)
                 {
-                    Debug.LogWarning($"Destroyed a duplicate {gameObject.name.Colorize(Color.magenta)} ({gameObject.GetInstanceID()}).");
+                    Debug.LogWarning($"Destroyed a duplicate {gameObject.name.Colorize(Palette.GoldenAmber)} ({gameObject.GetInstanceID()}).");
                     Destroy(gameObject);
                     return;
                 }